reoganized setup and test files
