+++ .github/workflows/run_tests_pr.yml
@@ -16,7 +16,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install flake8 pytest
-          if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+          pip install .[tables]
       - name: Test with pytest
         run: |
           pytest tests/
