+++ .github/workflows/run_tests_push.yml
@@ -1,7 +1,7 @@
 name: Run Tests on push
 on: 
   push:
-    branches: [ main, tables ]
+    branches: [ main ]
 jobs:
   test:
     runs-on: ubuntu-latest
